backend/requirements.txt
langchain
langgraph
langchain-openai
langchain-core
langchain-mcp-adapters
fastapi
uvicorn
requests
python-dotenv
loguru
